fixing some missing security checking

+++ app/api/brands/deleteBrand.php
@@ -1,7 +1,12 @@
 <?php
 require_once($_SERVER['DOCUMENT_ROOT'] . '/CarLog/app/controllers/brandController.php');
 
-$id = isset($_POST['brandId']) ? $_POST['brandId'] : null;
-$brandController = new BrandController();
+if ($_SERVER['USER']['role'] == 'admin') {
+    $id = isset($_POST['brandId']) ? $_POST['brandId'] : null;
+    $brandController = new BrandController();
+    $response = $brandController->deleteBrand($id);
+} else {
+    echo "You are not authorized to perform this action";
+}
 
-$response = $brandController->deleteBrand($id);
+header("Location: /CarLog/admin/manageBrandsPage/");
