+++ app/api/news/deleteNews.php
@@ -1,7 +1,10 @@
 <?php
 require_once($_SERVER['DOCUMENT_ROOT'] . '/CarLog/app/controllers/newsController.php');
 
-$id = isset($_POST['newsId']) ? $_POST['newsId'] : null;
-
-$newsController = new NewsController();
-$response = $newsController->deleteNews($id);
+if ($_SERVER['USER']['role'] == 'admin') {
+    $id = isset($_POST['newsId']) ? $_POST['newsId'] : null;
+    $newsController = new NewsController();
+    $response = $newsController->deleteNews($id);
+} else {
+    echo "You are not authorized to perform this action";
+}
